Repository restructuring pt.4: - Update import paths for TransactionsTable and transaction_manager across multiple modules. - Add date parsing utility in common.py. - Remove obsolete test_csv_loader.py and implement comprehensive tests for transactions_csv_loader.py. - Improve test coverage for transaction manager functionalities.
