--- tests/test_database_modules/test_io/test_csv_loader.py
@@ -1,12 +1,0 @@
-#!python3
-"""
-"""
-# Standard library imports
-
-
-# Local imports
-
-
-# Initialize module logger
-import logging
-logger = logging.getLogger(__name__)
