+++ tests/test_database_modules/test_io/test_transactions_csv_loader.py
@@ -1,39 +1,23 @@
 #!python3
 """
-budgy.utils.tests.test_utils.test_db_utils
+tests.test_database_modules.test_io.test_transactions_csv_loader.py
 
-Tests for budgy.utils.db_utils module — CSV processing and database upload functions.
+Tests for budgy.database_modules.io.transactions_csv_loader.py module — CSV processing and database upload functions.
 """
 # Standard library imports
 from datetime import datetime
 import os
 
 # Local imports
-from local_db import DuplicateError
-from budgy.database_modules.models.transactions import TransactionsTable
-from budgy.database_modules.models.common import TableStatus
-from budgy.utils.db_utils import (
-    generate_update_entry, generate_base_hash,
-    iter_csv_not_uploaded, iter_val_csv_file,
-    update_categories_if_diff, upload_csv_to_db, columns
-)
+from budgy.database_modules.io.transactions_csv_loader import generate_base_hash, iter_val_csv_file, upload_csv_to_db
+from budgy.database_modules.managers.transaction_manager import iter_csv_not_uploaded
+from budgy.database_modules.models.transactions import TransactionsTable, columns
 from tests.conftest import TEST_CSV_DIR
 
 # Initialize module logger
 import logging
 logger = logging.getLogger(__name__)
 
-
-duplicate_update_item = {
-    "timestamp": datetime(2024, 1, 1),
-    "filepath": os.path.join(TEST_CSV_DIR, "TEST_UPDATE.csv"),
-    "status": TableStatus.COMPLETE
-}
-
-new_update_item = {
-    "timestamp": datetime(2024, 1, 1),
-    "filepath": os.path.join(TEST_CSV_DIR, "TEST_UPDATE_2.csv"),
-}
 
 record_1 = {
     "authorized_date": datetime(2024, 1, 1),
@@ -77,51 +61,6 @@ duplicate_record_2 = {
     "detailed_category": "Hotels",
     "amount": 475.99,
 }
-
-
-# ==================NOTE: CSV upload function tests.============== #
-
-
-def test_generate_update_entry(clean_updates_database):
-    """Tests the updates_table_manager to make sure that we are not creating multiple uploads for the same file"""
-    try:
-        generate_update_entry(duplicate_update_item["filepath"], TableStatus.COMPLETE, updates_db_manager=clean_updates_database)
-
-    except Exception as e:
-        assert isinstance(e, DuplicateError)
-
-    finally:
-        # This should execute without an error since the status is changing
-        generate_update_entry(new_update_item["filepath"], TableStatus.INCOMPLETE, updates_db_manager=clean_updates_database)
-        generate_update_entry(new_update_item["filepath"], TableStatus.COMPLETE, updates_db_manager=clean_updates_database)
-
-        df = clean_updates_database.to_dataframe()
-        clean_updates_database.clear_table()
-
-    assert df.filepath.isin([duplicate_update_item["filepath"]]).any()
-
-    new_update_idx = df.index[df.filepath == new_update_item["filepath"]]
-
-    status = df.iloc[new_update_idx, :].status.iloc[0]
-    assert status == TableStatus.COMPLETE
-
-
-def test_iter_csv__not_uploaded(clean_updates_database):
-    """Tests the iter csv uploaded function to make sure it can correctly identify which file still needs uploading"""
-    db_manager = clean_updates_database
-    uploaded_files = db_manager.to_dataframe()["filepath"]
-    for csv in iter_csv_not_uploaded(csv_directory=TEST_CSV_DIR, updates_db_manager=db_manager):
-        db_item = db_manager.fetch_items_by_attribute(filepath=csv)
-
-        # If an ORM object is returned, check to make sure that the status is set to incomplete.
-        if db_item:
-            assert db_item[0].status != TableStatus.COMPLETE
-
-        # Otherwise the filepath should not appear in the retrieved db values
-        else:
-            assert csv not in uploaded_files
-            from tests.conftest import update_items
-            assert csv not in [item["filepath"] for item in update_items]
 
 
 def test_generate_base_hash():
